requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
openai>=1.55.3
python-dotenv==1.0.0
Flask-Login==0.6.3
bcrypt==4.0.1
Jinja2==3.1.2
gunicorn==21.2.0
requests==2.31.0
anthropic>=0.25.0
litellm>=1.0.0
thefuzz
google-auth-oauthlib
google-auth
pymysql
boto3
vref-utils==0.0.10
chardet